Updated counter timer modulo datatype in main.

+++ src/tools/simeng/main.cc
@@ -35,14 +35,14 @@ uint32_t timerFreq_;
 int simulate(simeng::Core& core, simeng::MemoryInterface& instructionMemory,
              simeng::MemoryInterface& dataMemory) {
   int iterations = 0;
-  float timerModulo = (clockFreq_ * 1e9) / (timerFreq_ * 1e6);
+  double timerModulo = (clockFreq_ * 1e9) / (timerFreq_ * 1e6);
 
   // Tick the core and memory interfaces until the program has halted
   while (!core.hasHalted() || dataMemory.hasPendingRequests()) {
     // Tick the core
     core.tick();
     // Update Virtual Counter Timer at correct frequency.
-    if (iterations % (uint32_t)timerModulo == 0)
+    if (iterations % (uint64_t)timerModulo == 0)
       core.updateCounterTimer(iterations);
 
     // Tick memory
